+++ pkg/config/config.go
@@ -59,6 +59,7 @@ type GCSConfig struct {
 	CompressionFormat string `yaml:"compression_format" envconfig:"GCS_COMPRESSION_FORMAT"`
 	Debug             bool   `yaml:"debug" envconfig:"GCS_DEBUG"`
 	Endpoint          string `yaml:"endpoint" envconfig:"GCS_ENDPOINT"`
+	StorageClass      string `yaml:"storage_class" envconfig:"GCS_STORAGE_CLASS"`
 }
 
 // AzureBlobConfig - Azure Blob settings section
@@ -390,6 +391,7 @@ func DefaultConfig() *Config {
 		GCS: GCSConfig{
 			CompressionLevel:  1,
 			CompressionFormat: "tar",
+			StorageClass:      "STANDARD",
 		},
 		COS: COSConfig{
 			RowURL:            "",