+++ ChangeLog.md
@@ -1,3 +1,8 @@
+# v1.6.0
+IMPROVEMENTS
+- add `storage_class` (GCS_STORAGE_CLASS) support for `remote_storage: gcs` fix [502](https://github.com/AlexAkulov/clickhouse-backup/issues/502)
+- upgrade aws golang sdk and gcp golang sdk to latest versions
+
 # v1.5.2
 IMPROVEMENTS
 - switch to go 1.19